Problem: image render calls are excessive
It is important to update the structure preview
image and column images to reflect changes in the
db, but updating from multiple places in the code (
and in response to events such as redis keyspace events)
causes several of the same call in short period of
time. This can be tested with fairytale using the
--structure-stagger-delay argument

Solution: use kivy's clock to schedule the call within a
short period of time rather than calling directly. Check
if an identical call has already been scheduled, and if so,
do not schedule another until cleared by the clock. The
checking of call equality is done using repr strings which is
messy. Attention should also be given to ways of speeding up /
caching aspects of the image rendering calls.