Problem: cannot sparsely fill expected and found
If rough amounts are defined for various possibilites
it would be useful to see the entire expected amount
filled in and updating with matched sources and handling
duplicates and more than expected amounts.

Solution: write the sparse_expected sorting code that
happens before image generation. This was done in a crude
way and there is much room for improvement. The gui already
contains a sparse_expected checkbox to toggle this view.