Problem: viewviewers are not modular
viewviewers are widgets that work with a the list
of dictionaries that each fold column consists of.
It is useful to provide a large degree of flexibility
since it allows for a wide range of interactions with
fold contents and is easily supported since any collection
of widgets can be added within the fold. Configuration
is handled by another widget class that provides a
constructor to the fold as it processes the dictionaries.

Solution: prototype a system of classes with a single
widget dynamically collecting and displaying in tabs the
configuration classes that provide an interface to
preconfiguring available viewviewer classes, which are
then made avaiable as constructors to folds. Improvements
might involve custom composites of viewviewer classes added
as a single widget.