Problem: folds are miniscule as columns increase
With more sources and columns, the fold width shrinks
and spaces between folds become too small to expand
beyond a few  pixels.

Solution: Display only a range of columns(a window)
in the folds. Select the range by clicking on the
preview image, which calulates the center column
clicked  and then requests an update of the folds.