Problem: folds cannot be noncontiguous columns
useful for view a view selected columns at once
or creating selections of columns according to
some criteria

Solution: nonlinear columns are a list of columns
that can be appended using the the right button on
the preview image. Folds tab will conists of these
columns, the list can be cleared by clicking the
left button on the preview image which will display
contiguous columns according to the column window span.