Problem: no way to run scripts on sources
useful to conditionally run some action or modify
field values of an added source. Initial use case
is to rotate binary keys using the existing programs
contained in machinic-image.

Solution: create a simple dsl called keyling to
write simple scripts that the ui can call on a
source or sources.